% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqgmt.R
\name{qqgmt}
\alias{qqgmt}
\title{qqgmt}
\usage{
qqgmt(gene_file, gene_set_name, description)
}
\arguments{
\item{gene_file}{a list of your interest genes}

\item{gene_set_name}{your interest gene set name}

\item{description}{whatever}
}
\value{
your interest gmt file
}
\description{
qqgmt
}
\examples{
qqgmt("2022minor_genes.xlsx", "Minor ZGA genes(Chunxia,2022)", "NA")
}
